Added a CurlCommandLine to allow the jar to be run from command line.
Created CurlCommandLine to allow the jar to be ran using `java -jar curl.jar <curl options>`.
Also used the maven-shade-plugin to add all of the dependencies to into a single jar file to
allow this to be used in locations where only java is installed.

The current build was 3.3M, which isn't too bad.
May be able to remove more but would have to somehow do some dynamic analysis
to determine all of the required classes by the dependencies. Or looking into how
minimizeJar works.